+++ src/main/java/ru/practicum/shareit/comment/dto/CommentDto.java
@@ -1,0 +1,19 @@
+package ru.practicum.shareit.comment.dto;
+
+import lombok.Data;
+
+import java.time.LocalDateTime;
+
+@Data
+public class CommentDto {
+
+    private Long id;
+
+    private Long item;
+
+    private String text;
+
+    private String authorName;
+
+    private LocalDateTime created;
+}
