+++ src/main/java/ru/practicum/shareit/comment/mapper/CommentMapper.java
@@ -1,18 +1,16 @@
-package ru.practicum.shareit.item.mapper;
+package ru.practicum.shareit.comment.mapper;
 
-import ru.practicum.shareit.item.dto.CommentCreateDto;
-import ru.practicum.shareit.item.dto.CommentDto;
-import ru.practicum.shareit.item.model.Comment;
-import ru.practicum.shareit.user.mapper.UserMapper;
+import ru.practicum.shareit.comment.dto.CommentCreateDto;
+import ru.practicum.shareit.comment.dto.CommentDto;
+import ru.practicum.shareit.comment.model.Comment;
 
 public final class CommentMapper {
 
     public static CommentDto toCommentDto(Comment comment) {
         CommentDto commentDto = new CommentDto();
         commentDto.setId(comment.getId());
+        commentDto.setItem(comment.getItem().getId());
         commentDto.setText(comment.getText());
-        commentDto.setItem(comment.getItem() != null ? ItemMapper.toItemDto(comment.getItem()) : null);
-        commentDto.setAuthor(comment.getAuthor() != null ? UserMapper.toUserDto(comment.getAuthor()) : null);
         commentDto.setCreated(comment.getCreated());
         return commentDto;
     }
@@ -21,8 +19,6 @@ public final class CommentMapper {
         Comment comment = new Comment();
         comment.setId(commentDto.getId());
         comment.setText(commentDto.getText());
-        comment.setItem(commentDto.getItem() != null ? ItemMapper.toItem(commentDto.getItem()) : null);
-        comment.setAuthor(commentDto.getAuthor() != null ? UserMapper.toUser(commentDto.getAuthor()) : null);
         comment.setCreated(commentDto.getCreated());
         return comment;
     }
