+++ src/main/java/ru/practicum/shareit/item/ItemController.java
@@ -6,6 +6,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+import ru.practicum.shareit.comment.dto.CommentCreateDto;
+import ru.practicum.shareit.comment.dto.CommentDto;
 import ru.practicum.shareit.item.dto.*;
 import ru.practicum.shareit.item.service.ItemService;
 
@@ -59,8 +61,8 @@ public class ItemController {
     }
 
     @GetMapping
-    public Collection<ItemWithBookingDto> getItems(@RequestHeader(name = "X-Sharer-User-Id", required = false, defaultValue = "-1")
-                                                   @PositiveOrZero Long userId) {
+    public Collection<ItemDto> getItems(@RequestHeader(name = "X-Sharer-User-Id", required = false, defaultValue = "-1")
+                                        @PositiveOrZero Long userId) {
         log.info("Received a request to get a items with user id: {}", userId);
         return itemService.getItems(userId);
     }
