+++ src/main/java/ru/practicum/shareit/item/dto/ItemDto.java
@@ -1,6 +1,8 @@
 package ru.practicum.shareit.item.dto;
 
 import lombok.Data;
+import ru.practicum.shareit.booking.dto.BookingShortDto;
+import ru.practicum.shareit.comment.dto.CommentDto;
 import ru.practicum.shareit.request.dto.ItemRequestDto;
 import ru.practicum.shareit.user.dto.UserDto;
 
@@ -14,5 +16,7 @@ public class ItemDto {
     private Boolean available;
     private UserDto owner;
     private ItemRequestDto request;
+    private BookingShortDto lastBooking;
+    private BookingShortDto nextBooking;
     private List<CommentDto> comments;
 }
