--- src/main/java/ru/practicum/shareit/item/dto/ItemWithBookingDto.java
@@ -1,17 +1,0 @@
-package ru.practicum.shareit.item.dto;
-
-import lombok.Data;
-import ru.practicum.shareit.booking.dto.BookingShortDto;
-
-import java.util.List;
-
-@Data
-public class ItemWithBookingDto {
-    private Long id;
-    private String name;
-    private String description;
-    private Boolean available;
-    private BookingShortDto lastBooking;
-    private BookingShortDto nextBooking;
-    private List<CommentDto> comments;
-}
