+++ src/main/resources/schema.sql
@@ -47,42 +47,6 @@ CREATE TABLE IF NOT EXISTS items (
 );
 
 
--- public.bookings definition
-
--- Drop table
-
--- DROP TABLE bookings;
-
-CREATE TABLE IF NOT EXISTS bookings (
-	id int4 NOT NULL GENERATED BY DEFAULT AS IDENTITY,
-	start_date timestamp NULL,
-	end_date timestamp NULL,
-	item_id int4 NOT NULL,
-	booker_id int4 NULL,
-	status varchar(15) NULL,
-	CONSTRAINT bookings_pk PRIMARY KEY (id),
-	CONSTRAINT bookings_start_end_check CHECK ((start_date < end_date)),
-	CONSTRAINT bookings_items_fk FOREIGN KEY (item_id) REFERENCES items(id),
-	CONSTRAINT bookings_users_fk FOREIGN KEY (booker_id) REFERENCES users(id)
-);
-
-
--- public."comments" definition
-
--- Drop table
-
--- DROP TABLE "comments";
-
-CREATE TABLE IF NOT EXISTS "comments" (
-	id int4 NOT NULL GENERATED BY DEFAULT AS IDENTITY,
-	"text" varchar(250) NOT NULL,
-	item_id int4 NOT NULL,
-	author_id int4 NOT NULL,
-	CONSTRAINT comments_pk PRIMARY KEY (id),
-	CONSTRAINT comments_items_fk FOREIGN KEY (item_id) REFERENCES items(id),
-	CONSTRAINT comments_users_fk FOREIGN KEY (author_id) REFERENCES users(id)
-);
-
 -- public."comments" definition
 
 -- Drop table
